refactor(day-diet): encapsulate store signals and extract realtime logic
