+++ src/modules/diet/day-diet/infrastructure/supabase/supabaseDayRepository.ts
@@ -13,10 +13,7 @@ import {
   createErrorHandler,
   wrapErrorWithStack,
 } from '~/shared/error/errorHandler'
-import {
-  registerSubapabaseRealtimeCallback,
-  supabase,
-} from '~/shared/utils/supabase'
+import { supabase } from '~/shared/utils/supabase'
 
 export const SUPABASE_TABLE_DAYS = 'days'
 
@@ -31,51 +28,6 @@ export function createSupabaseDayRepository(): DayRepository {
     updateDayDiet,
     deleteDayDiet,
   }
-}
-
-/**
- * Sets up granular realtime subscription for day diet changes
- * @param onDayDietChange - Callback for granular updates with event details
- */
-export function setupDayDietRealtimeSubscription(
-  onDayDietChange: (event: {
-    eventType: 'INSERT' | 'UPDATE' | 'DELETE'
-    old?: DayDiet
-    new?: DayDiet
-  }) => void,
-): void {
-  registerSubapabaseRealtimeCallback(
-    SUPABASE_TABLE_DAYS,
-    (payload: unknown) => {
-      // eslint-disable-next-line @typescript-eslint/consistent-type-assertions
-      const payloadData = payload as {
-        eventType?: string
-        old?: unknown
-        new?: unknown
-      }
-
-      // eslint-disable-next-line @typescript-eslint/consistent-type-assertions
-      const eventType = payloadData.eventType as 'INSERT' | 'UPDATE' | 'DELETE'
-
-      // Parse old and new records if available
-      const oldRecord =
-        payloadData.old !== null
-          ? dayDietSchema.safeParse(payloadData.old)
-          : null
-      const newRecord =
-        payloadData.new !== null
-          ? dayDietSchema.safeParse(payloadData.new)
-          : null
-
-      onDayDietChange({
-        eventType,
-        old:
-          oldRecord !== null && oldRecord.success ? oldRecord.data : undefined,
-        new:
-          newRecord !== null && newRecord.success ? newRecord.data : undefined,
-      })
-    },
-  )
 }
 
 // TODO:   better error handling
