+++ src/modules/diet/meal/application/meal.ts
@@ -1,5 +1,5 @@
+import { currentDayDiet } from '~/modules/diet/day-diet/application/dayDietUseCases'
 import { updateDayDiet } from '~/modules/diet/day-diet/application/dayDietUseCases'
-import { currentDayDiet } from '~/modules/diet/day-diet/application/dayDietStore'
 import { demoteNewDayDiet } from '~/modules/diet/day-diet/domain/dayDiet'
 import { updateMealInDayDiet } from '~/modules/diet/day-diet/domain/dayDietOperations'
 import { type Meal } from '~/modules/diet/meal/domain/meal'
